+++ package.json
@@ -3,8 +3,8 @@
   "version": "3.2.0",
   "description": "Use Amazon Athena with Grafana",
   "scripts": {
-    "build": "webpack -c ./.config/webpack/webpack.config.ts --env production",
-    "dev": "webpack -w -c ./.config/webpack/webpack.config.ts --env development",
+    "build": "webpack -c ./webpack.config.ts --env production",
+    "dev": "webpack -w -c ./webpack.config.ts --env development",
     "e2e": "playwright test",
     "lint": "eslint --cache .",
     "lint:fix": "npm run lint -- --fix && prettier --write --list-different .",
@@ -16,7 +16,7 @@
     "test:ci": "jest --passWithNoTests --maxWorkers 4",
     "typecheck": "tsc --noEmit",
     "vulncheck": "go run golang.org/x/vuln/cmd/govulncheck@v1.3.0 ./...",
-    "watch": "webpack -w -c ./.config/webpack/webpack.config.ts --env development"
+    "watch": "webpack -w -c ./webpack.config.ts --env development"
   },
   "author": "Grafana Labs",
   "license": "Apache-2.0",
@@ -90,7 +90,7 @@
     "terser-webpack-plugin": "^5.3.14",
     "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
-    "typescript": "5.9.3",
+    "typescript": "6.0.3",
     "webpack": "^5.102.0",
     "webpack-cli": "^6.0.1",
     "webpack-livereload-plugin": "^3.0.2",
