+++ src/AnnotationQueryEditor.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { AthenaQuery, AthenaDataSourceOptions } from './types';
 import { QueryEditorProps } from '@grafana/data';
-import { DataSource } from 'datasource';
-import { QueryEditorForm } from './QueryEditorForm';
+import { DataSource } from '@/datasource';
+import { QueryEditorForm } from '@/QueryEditorForm';
+import type { AthenaQuery, AthenaDataSourceOptions } from '@/types';
 
 export function AnnotationQueryEditor(props: QueryEditorProps<DataSource, AthenaQuery, AthenaDataSourceOptions>) {
   return <QueryEditorForm {...props} hideOptions />;
