+++ src/VariableQueryEditor.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { AthenaQuery, AthenaDataSourceOptions } from './types';
+import { AthenaQuery, AthenaDataSourceOptions } from '@/types';
 import { QueryEditorProps } from '@grafana/data';
-import { DataSource } from 'datasource';
-import { QueryEditorForm } from './QueryEditorForm';
+import { DataSource } from '@/datasource';
+import { QueryEditorForm } from '@/QueryEditorForm';
 
 export function VariableQueryCodeEditor(props: QueryEditorProps<DataSource, AthenaQuery, AthenaDataSourceOptions>) {
   return <QueryEditorForm {...props} hideOptions />;
