+++ src/annotationSupport.ts
@@ -1,4 +1,4 @@
-import { AnnotationQueryEditor } from './AnnotationQueryEditor';
+import { AnnotationQueryEditor } from '@/AnnotationQueryEditor';
 
 export const annotationSupport = {
   QueryEditor: AnnotationQueryEditor,
