+++ src/language/completionItemProvider.ts
@@ -5,7 +5,7 @@ import {
   TableDefinition,
   TableIdentifier,
 } from '@grafana/plugin-ui';
-import { MACROS } from './macros';
+import { MACROS } from '@/language/macros';
 
 interface CompletionProviderGetterArgs {
   getTables: (d: TableIdentifier | null) => Promise<TableDefinition[]>;
@@ -14,17 +14,17 @@ interface CompletionProviderGetterArgs {
 
 export const getAthenaCompletionProvider: (args: CompletionProviderGetterArgs) => LanguageCompletionProvider =
   ({ getTables, getColumns }) =>
-  (monaco, language) => {
-    return {
-      // get standard SQL completion provider which will resolve functions and macros
-      ...(language && getStandardSQLCompletionProvider(monaco, language)),
-      triggerCharacters: ['.', ' ', '$', ',', '(', "'"],
-      tables: {
-        resolve: getTables,
-      },
-      columns: {
-        resolve: (t?: TableIdentifier) => getColumns(t?.table ?? ''),
-      },
-      supportedMacros: () => MACROS,
+    (monaco, language) => {
+      return {
+        // get standard SQL completion provider which will resolve functions and macros
+        ...(language && getStandardSQLCompletionProvider(monaco, language)),
+        triggerCharacters: ['.', ' ', '$', ',', '(', "'"],
+        tables: {
+          resolve: getTables,
+        },
+        columns: {
+          resolve: (t?: TableIdentifier) => getColumns(t?.table ?? ''),
+        },
+        supportedMacros: () => MACROS,
+      };
     };
-  };
