+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export * from "@/types/config.types"
+export * from "@/types/query.types";
