+++ tests/e2e/annotationsEditor.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from '@grafana/plugin-e2e';
-import { selectors } from '../../src/tests/selectors';
+import { selectors } from '@/selectors/selectors';
 
 test('should render annotations editor', async ({ annotationEditPage, page }) => {
   await annotationEditPage.datasource.set('AWS Athena');
