+++ tsconfig.json
@@ -1,3 +1,11 @@
 {
-  "extends": "./.config/tsconfig.json"
+  "extends": "./.config/tsconfig.json",
+  "compilerOptions": {
+    "ignoreDeprecations": "6.0",
+    "rootDir": ".",
+    "paths": {
+      "@/*": ["./src/*"],
+    },
+  },
+   "include": ["./src", "./tests", "./.config/types"]
 }
